refactor: PR 리뷰 피드백 적용(#37)
- post에서 record로 기록 통일
