+++ frontend/src/app/(map)/_components/PinMarker.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { AdvancedMarker } from '@vis.gl/react-google-maps';
 import type { Marker } from '@googlemaps/markerclusterer';
-import type { PostListItem } from '@/lib/types/post';
+import type { PostListItem } from '@/lib/types/record';
 import Image from 'next/image';
 
 export type PinMarkerProps = {
