+++ frontend/src/app/(post)/record/_components/RecordDetailHeaderActions.tsx
@@ -8,7 +8,7 @@ import {
   DrawerTitle,
 } from '@/components/ui/drawer';
 import { Popover } from '@/components/ui/popover';
-import { MemoryRecord } from '@/lib/types/post';
+import { MemoryRecord } from '@/lib/types/record';
 import {
   PopoverClose,
   PopoverContent,
